refactor: relocate MTD diarization responsibilities
迁移 MTD 说话人分离职责

将通用模型接口收敛为：
decode_snapshot(request_id, pcm16, sample_rate, is_final)

OfficialMtdClient 内部构建注册音频布局、生成 decoder prefix 并维护 exemplar pool；MultiSpeakerTurnContextManager 负责通用音频缓冲、partial/final 调度、事件发布和 ASR 结果合并。删除 MtdDiarizationManager，将通用调度配置迁移至 service_config.multi_speaker.diarization，并补充模型与 serving 回归测试。
